# ⚡ perf: Add a tail call dispatch evm implementation