… bench/interpreter_comparison_benchmark.zig
… bench/official/evms/zig/src/main.zig
… bench/official/results.md
… bench/official/src/Orchestrator.zig
… bench/official/src/evm-runner-call2.zig
… bench/official/src/main.zig
… bench/zbench_runner.zig
… src/evm/evm/call_contract.zig
… src/evm/evm/interpret2.zig
… src/evm/evm/interpret2_simple.zig
… src/evm/evm/tailcall_wrappers.zig
… src/evm/evm/tailcalls.zig
… src/evm/evm/threadify.zig
… src/evm/execution/control.zig
… src/evm/execution/system.zig
… src/evm/frame.zig
… src/evm/instruction_tag_test.zig
… src/evm/root.zig
… src/evm/size_buckets.zig
… src/evm/state/state.zig
… src/evm/tailcall_execution_func.zig
… test/differential/call2_differential_test.zig
… test/differential/comparison_differential_test.zig
… test/differential/erc20_differential_test.zig
… test/differential/package.zig
… test/differential/selfdestruct_edge_cases_differential_test.zig
… test/differential/ten_thousand_hashes_differential_test.zig
… test/evm/interpret2_comprehensive_test.zig
… test/evm/interpret2_simple_test.zig
… test/evm/interpret2_test.zig