## Summary
- Replaced O(n) bytecode scanning with O(1) precomputed mappings
- Significant performance improvements for call2 interpreter  
- Maintains compatibility with existing code

## Problem
The call2 interpreter (using tailcall dispatch) was performing O(n) bytecode scanning on every PUSH, JUMP, JUMPI, and PC operation. This created quadratic complexity for contracts with many instructions.

## Solution
Created a `SimpleAnalysis` module that precomputes:
- `inst_to_pc`: Maps instruction index → PC value
- `pc_to_inst`: Maps PC → instruction index
- `push_values`: HashMap of PC → push value

These mappings enable O(1) lookups in critical opcodes.

## Performance Results
**ten-thousand-hashes benchmark:**
- Before: ~30ms
- After: ~4ms  
- **~7x faster**

**erc20-transfer benchmark:**
- Before: ~45ms
- After: ~7ms
- **~6x faster**

The call2 interpreter is now competitive with the standard interpreter.

## Changes
1. ✨ Added `SimpleAnalysis` module for precomputed mappings
2. ♻️ Updated Frame struct to store analysis pointer
3. ♻️ Integrated SimpleAnalysis into interpret2
4. ⚡ Optimized tailcall opcodes with cached lookups
5. 🔧 Updated evm-runner-call2 import path

## Test Plan
- [x] All existing tests pass
- [x] Benchmarks show significant performance improvement
- [x] No regressions in functionality

🤖 Generated with [Claude Code](https://claude.ai/code)